Add css fixes
