+++ blocks/cards/cards.js
@@ -1,5 +1,3 @@
-import { createOptimizedPicture } from '../../scripts/aem.js';
-
 export default function decorate(block) {
   /* change to ul, li */
   const ul = document.createElement('ul');
@@ -9,10 +7,23 @@ export default function decorate(block) {
     [...li.children].forEach((div) => {
       if (div.children.length === 1 && div.querySelector('picture')) div.className = 'cards-card-image';
       else div.className = 'cards-card-body';
+
+      const textWrapper = document.createElement('div');
+      textWrapper.append(...div.children);
+      div.append(textWrapper);
+      const pic = div.querySelector('picture');
+
+      if (pic) {
+        const el = document.createElement('div');
+        el.classList.add('picture-wrapper');
+        pic.parentElement.remove();
+        el.append(pic);
+        div.prepend(el);
+      }
     });
     ul.append(li);
   });
-  ul.querySelectorAll('img').forEach((img) => img.closest('picture').replaceWith(createOptimizedPicture(img.src, img.alt, false, [{ width: '750' }])));
+
   block.textContent = '';
   block.append(ul);
 }
