+++ blocks/footer/footer.css
@@ -6,7 +6,7 @@ footer {
 
 footer .footer {
   max-width: 1200px;
-  margin: right;
+  margin: auto;
 }
 
 footer .footer p {
