+++ styles/styles.css
@@ -262,3 +262,11 @@ h2 {
       line-height: 2.75rem;
   }
 }
+
+.section.gray-bg {
+  background-color: #f7f7f7;
+}
+
+.block a.button {
+  padding: 0;
+}
